Modified invert button behavior.
Recalculates "shift" value based on the value of the invert button.
If invert is true, it sets y shift to -ymax clipping value
if false, y shift is set to ymin clipping values

The problem with this is that only works for pixels. For x,y data that
is not pixels, it goofs it. In the next fix, this will change to max and
min y value data, not index (pixel) data.